feat: 오, 육단계 기능 추가 refactor: Reservation 필드 (date, time) 분리

+++ src/main/java/roomescape/controller/ReservationController.java
@@ -49,7 +49,7 @@ public class ReservationController {
   }
 
   @DeleteMapping("/reservations/{id}")
-  public ResponseEntity<Void> createReservation (@PathVariable Integer id) {
+  public ResponseEntity<Void> createReservation (@PathVariable Long id) {
 
     Reservation reservation = reservationService.delete(id);
     return ResponseEntity.noContent().build();
